Conflicts:
	newscoop/admin-files/articles/comments/show_comments.php
	newscoop/admin-files/comments/index.php
	newscoop/admin-style/admin_stylesheet.css
	newscoop/admin-style/form.css
	newscoop/application/layouts/scripts/admin.phtml
	newscoop/bin/cli_script_lib.php
	newscoop/install/sql/upgrade/3.5.x/data-required.sql
	newscoop/install/sql/upgrade/3.5.x/tables.sql
	newscoop/js/tinymce/plugins/campsiteimage/popup.php
	newscoop/library/Newscoop/Entity/Article.php
	newscoop/library/Newscoop/Entity/Proxy/NewscoopEntityPublicationProxy.php
	newscoop/library/Newscoop/Entity/Publication.php

+++ newscoop/application/modules/admin/controllers/StaffController.php
@@ -93,7 +93,7 @@ class Admin_StaffController extends Zend_Controller_Action
 
         $this->view->actions = array(
             array(
-                'label' => getGS('Edit access'),
+                'label' => getGS('Permissions'),
                 'module' => 'admin',
                 'controller' => 'staff',
                 'action' => 'edit-access',
@@ -123,14 +123,24 @@ class Admin_StaffController extends Zend_Controller_Action
     public function deleteAction()
     {
         $this->_helper->acl->check('user', 'delete');
-
+        
         $staff = $this->_helper->entity->get(new Staff, 'user');
-        $this->repository->delete($staff);
-
-        $this->_helper->entity->getManager()->flush();
-
-        $this->_helper->flashMessenger(getGS('Staff member deleted.'));
-        $this->_helper->redirector->gotoSimple('index');
+        
+        if (Zend_Auth::getInstance()->getIdentity() == $staff->getId()) $permitted = false;
+        else $permitted = true;
+        
+        if ($permitted) {
+            $this->repository->delete($staff);
+    
+            $this->_helper->entity->getManager()->flush();
+    
+            $this->_helper->flashMessenger(getGS('Staff member deleted.'));
+            $this->_helper->redirector->gotoSimple('index');
+        }
+        else {
+            $this->_helper->flashMessenger(getGS('Self-delete is not permitted.')); // should be translateable
+            $this->_helper->redirector->gotoSimple('index');
+        }
     }
 
     public function tableAction()
@@ -188,6 +198,7 @@ class Admin_StaffController extends Zend_Controller_Action
                 'action' => 'add',
                 'resource' => 'user',
                 'privilege' => 'manage',
+                'class' => 'add',
             ),
         );
     }
